Error Handling

+++ src/articles/articles.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import { CreateArticleDto } from './dto/create-article.dto';
 import { UpdateArticleDto } from './dto/update-article.dto';
+import { Injectable, NotFoundException } from '@nestjs/common';
 
 @Injectable()
 export class ArticlesService {
@@ -23,14 +23,22 @@ export class ArticlesService {
     return this.prisma.article.findUnique({ where: { id } });
   }
 
-  update(id: number, updateArticleDto: UpdateArticleDto) {
+  async update(id: number, updateArticleDto: UpdateArticleDto) {
+    const article = await this.findOne(id);
+
+    if (!article) throw new NotFoundException(`Article with ${id} does not exist.`);
+
     return this.prisma.article.update({
       where: { id },
       data: updateArticleDto,
     });
   }
 
-  remove(id: number) {
+  async remove(id: number) {
+    const article = await this.findOne(id);
+
+    if (!article) throw new NotFoundException(`Article with ${id} does not exist.`);
+
     return this.prisma.article.delete({ where: { id } });
   }
 }
